Deleted chrome/utility/services.cc patch
fix https://github.com/brave/brave-browser/issues/16457

Used BraveContentUtilityClient::RegisterMainThreadServices() instead.
