+++ chromium_src/chrome/utility/DEPS
@@ -1,6 +1,3 @@
 include_rules = [
   "+../../../../../chrome/utility/importer",
-  "+brave/common",
-  "+chrome/common",
-  "+chrome/utility",
 ]
